+++ lib/run/workers/ParticleJobs.h
@@ -54,6 +54,32 @@ public:
     virtual void evaluate(const RunSettings& global, IRunCallbacks& callbacks) override;
 };
 
+class MultiJoinParticlesJob : public IParticleJob {
+private:
+    int slotCnt = 3;
+    bool moveToCom = false;
+
+public:
+    MultiJoinParticlesJob(const std::string& name)
+        : IParticleJob(name) {}
+
+    virtual std::string className() const override {
+        return "multi join";
+    }
+
+    virtual UnorderedMap<std::string, JobType> getSlots() const override {
+        UnorderedMap<std::string, JobType> map;
+        for (int i = 0; i < slotCnt; ++i) {
+            map.insert("particles " + std::to_string(i + 1), JobType::PARTICLES);
+        }
+        return map;
+    }
+
+    virtual VirtualSettings getSettings() override;
+
+    virtual void evaluate(const RunSettings& global, IRunCallbacks& callbacks) override;
+};
+
 class TransformParticlesJob : public IParticleJob {
 private:
     struct {